+++ CHANGELOG.md
@@ -13,6 +13,7 @@ Additions
 - Added GenericFont SplitStringSeparate which differentiates between existing newline characters and splits due to maximum width
 - Added StaticSpriteBatch class
 - Added missing easing functions Quart and Quint to Easings
+- Added RotationVector extension methods for Matrix and Quaternion
 
 Improvements
 - Cache TokenizedString inner offsets for non-Left text alignments to improve performance